allow persisting backups in another location
Will by default attempt to use
  * /var/cache/openaps-ruination/

If the `BACKUP_AREA` environment variable is defined and points to a
directory, it will be used.  If any argument is given to the script, eg:

    oref0-reset-git ../backups

And the argument points to a directory, that location will be used
instead.

If none of those three locations works as a backup directory, we fall
back to using /tmp/, per old behavior.