Now carbs.js is able to handle the pump setting "exchanges" as well
I added a check for the carbratio.data == "exchanges" because this is the alternative setting to the unit "grams". If "exchanges" is used in the pump the information behind the ratio is how many insulin units are needed for 1 bread exchange unit (= 12 grams of carbohydrate). To make this entered value fitting in the Openaps system, which uses "grams" as an value that says how many grams are matching 1 unit of insulin, the carbRatio.ratio is recalculated to grams per unit in the case the pump value is based on insulin per bread exchange unit. For some users (like me) it could be better to adjust the recalculating factor from 12 to 10. This is the case if the German "Kohlenhydrateinheit" (= 10 grams of carbohydrate) is used instead of the classic bread exchange unit (= 12 grams of carbohydrate). First I was thinking about taking 11 as a mean value between this to options with a calculating error of 10 percent or even less, but then I thought that nobody in the community had a problem so far with the carb_ratio in "exchanges" and if someone has a problem like me, the possibility that he uses "Kohlenhydrateinheit" is even less. Therefor I decided to enter 12 and adjust it to 10 in my one version ;-)