allow missing --meal doc
Tries to retain previous behavior via arguments.
Introduces new `--meal` flag.
If meals is absent, a warning is generated, but processing
continues.
With --no-missing-meal-ok set, an error will be raised, and processing will stop.

Allows @jasoncalabrese, @scottleibrand and others with
flagging/failed data to allow loop to do as much as
possible. :-)