--no-missing-auto-sens-ok, make more flexible
Allows the tool to try harder, thanks to feedback from
@scottleibrand and @jasoncalabrese.

With `--missing-auto-sens-ok` (the default) set, if the
auto-sens data is missing, a warning will be issued, and
processing will proceed, allowing a fall-back mode for
missing data.  With the complimentary
`--no-missing-auto-sens-ok` set, if the auto-sens data is
missing an error will be generated and processing will
stop.