use --date dateString explicitly
Thanks to @johners97 for finding, and @cjo20 for pinpointing a bug
where uploading entries to Nightscout used to not work with
Medtronic data.  With Dexcom, the `select` function was defaulting
to look at the `display_time` attribute, this change makes it
explicit that when working with Nightscout entries, we want to use the
`dateString`.  This change should help incorporate Medtronic data a little more
easily.