Project UAM impact out to DIA hours
Now that we're excluding future temps from predBGs, they're not as much
"up and to the right" when we have a long zero temp running.  So it should
be safter to extend UAM impact out up to DIA hours when we're at peak recent
deviations.  As soon as deviations start to come down that'll limit the UAM
impact predictions.