Upload entries from G5 to nightscout (#718)
* Update packages.json

Add the new script

* oref0-upload-entries.json

This script uploads new CGM entries to Nightscout. If the network connection is lost for a period of time the entries from the last time the upload was successful (stored in ``entries-last-date.json``) buffer in ``entries-upload.json`` until the next successful upload, at which point the file is deleted. This allows the device to be offline for any number of days and you get all the entries, not just the last 288, once it comes back online.

* Rename oref0-upload-entries.json to oref0-upload-entries.sh

* Update oref0-setup.sh to support "g5-upload" CGM type

* Add the Bash indicator