Don't swallow errors in the pump loop.
It hides errors and makes troubleshooting harder.
Showing errors printed to stderr in the pump loop
logs will encourage having a more stable system.

Alternatively if there was a reason for this, it
could be made configurable, but for now i think
it's better to keep info printed to stderr