Fixed the dev oref0-pump-loop script to drop x12-incompaitble reports IF an x12 pump is detected
In previous versions of oref0, this action was done by manually implementing an alias-change (as documented in past versions of the oref0 code).

Since the code here is not done via an alias anymore, but rather spelled out in the script, an if/then/else block is required to remove the offending reports when an incompatible pump is detected. 

Tested and verified against my 712.