Fully reload pumphistory file when clock is changed.
This minimizes chances of missing records due to the way we query the pump
for new record since the last timestamp.

This doesn't fix all issues as other rigs might miss the records if clock is moved
back.

(last record at 1:00 so we query the pump for all the records r > 1:00. Clock is
moved from 1:05 to 00:05, and a bolus is given at 00:05.
Bolus won't be included by the pumphistory update because 00:05 <= 1:00)