Refactoring: Add and use file_is_recent function
Add file_is_recent and file_is_recent_and_min_size, replacing an idiom
with find |grep that there was a lot of throughout the codebase. There
is one significant change relative to that idiom: file_is_recent expects
the given filename to be an exact match, while the original idiom
required only a substring match, and in some cases this was actually
depended on (eg by leaving off the ".json" part of the filename). This
eliminates some very-minor bugs where if a file named something.json.new
existed its timestamp would be used even if something.json didn't exist
or was too old, eliminated the risk of filename name-substring clashes
and helps things pass the "grep test".

Conflicts:
	bin/oref0-ns-loop.sh
	bin/oref0-pump-loop.sh