Refactor: Add and use check_pref_bool, other preferences-file functions
This adds a collection of utility functions for getting things from the
preferences.json, and makes the bash scripts use them. Right now the
preferences file only gets a limited amount of use on the bash side, but
that will increase, since I'm going to try to migrate the data that's
stored in weird hard-to-reach places (like crontab) into the preferences
file.