Redo oref0-log-shortcuts
The script oref0-log-shortcuts adds a collection of aliases to the
user's .bash_profile, so that they can conveniently look at logs/edit
their configs/etc. Previously this was done by grepping for part of each
alias, and appending a line with the alias if it wasn't found. However,
this interacts badly with upgrades; if the user had already had their
.bash_profile modified and one of the aliases was changed in a version
upgrade, the change wouldn't be applied to their profile. It also meant
that the contents of the aliases themselves got partially repeated, with
some tricky quoting.

This script makes oref0-log-shortcuts instead modify bash_profile to
source itself, so that if the aliases it defines changes, the aliases
the user gets will also change with no further .bash_profile
modifications necessary. If run on a .bash_profile which was modified by
an older version of oref0, it will clear out any old aliases which it
recognizes as exact matches, so old script-generated aliases should
migrate properly while changes made by users should be left untouched.