Remove unused OpenAPS aliases
The JSON files in lib/oref0-setup contain a bunch of aliases for
openaps, which are short-to-medium-length shell scripts jammed into a
single line. The majority of them have already been replaced by
functions in oref0-pump-loop.sh and oref0-ns-loop.sh, but are still
hanging around. This removes the ones that are fully unused (without
touching the ones that had at least one usage left). It also modified
oref0-setup.sh so that aliases which no longer exist get removed.

My methodology for determining that an OpenAPS alias is unused:

 * Search the oref0 codebase for the alias name, or
   "openaps\s*aliasname" if that has too many false positives
 * Search the master branch of github.com/openaps/docs for "openaps aliasname"
   looking for suggestions that users run it (which would make it
   effectively a public API)
 * Search the master branch of openaps for aliasname

If after inspecting the results of those searches none of them look like
usages of the alias, I assume it's safe to remove. Due to recent work by
others migrating code from aliases to oref0-pump-loop.sh and
oref0-ns-loop.sh, this is sufficient to remove more than half of the
aliases (though not yet all).

I also added comments (as "//":"text" in JSON) for some of the remaining
aliases, indicating where their usages are.

Conflicts:
	bin/oref0-setup.sh