Convert "get-bg" alias into oref0-get-bg.sh
The get-bg alias is the entry point of a bunch of openaps aliases
calling each other, and was tricky to resolve because it's conditional
on CGM type. Because cron uses the process name to tell whether it's
currently running, it has to be a full script and can't just be a shell
function.