Improved incremental pump history update.
Querying by top record timestamp is sometimes racy because of different
reasons:
 - some records have the same timestamp such as Bolus and BolusWizard.
 - some records have timestamps out of order such as DailyTotal records.
 - Timestamps can also become non monotonically increasing whenn clock
has been changed.

To avoid all these problems, a new argument was added to the pumphistory
command to query from a given record id. The record id is the base64
encoded string of the record.

Note this is not available yet in the latest go pump libaries, so this
change depends on https://github.com/ecc1/medtronic/pull/8