Split the pre-bolus ESC key presses in two commands.
The 4 ESC seems to be done too quickly now that we moved to the go
library.
It seems like it isn't as good at avoiding A52 errors as what
was done on master. I think on master the pump is wakened up again
between each key presses, or the python layer adds some latency which
isn't there on the new version.

The compromise here is to send the first ESC, wait for 0.5s before
sending the remaining ESC.

From testing on a test pump, this seems to reduce the
chance of getting an A52 error.