hotfix mdt not syncing bug (#1197)
When monitor/cgm-mm-glucosedirty.json is empy (maybe caused by filesystem damage or sudden power loss), the script will end with a "timeinterval less than 5 minutes" massage. There is no way out of this loop until the file gets deleted. I've added a check for empty file and delete the file if it is empty to force the script to a full refresh.

Testen on my pi0 based rig. The error is easy to produce with "rm monitor/cgm-mm-glucosedirty.json && touch monitor/cgm-mm-glucosedirty.json" on the old scrit. With the new script it works fine.