Turn off screen only after 5 minutes.
Some times, (many times starting on first install) one can not connect the pi to WiFi, but can connect a screen.
This pr will give them 5 minutes before the screen is turned off, to allow them to either enable connection to WiFi, or disable screen turning off.

One more thing that is changes is that hat might be on spidev0.[01] so screen is shutdown in both cases. (without this change, on adafruit bonnet screen is not turned off at all).

Signed-off-by: Tzachi Dar <tzachi.dar@gmail.com>