Add support for bgproxy.
A new preference option is added enableEnliteBgproxy. When set to true
the bgproxy command will be call at the end of every loop sending the
glucose data to the medtronic pump.

This gives an option to setup low and high alarms directly from the
pumps and also see the glucose trend when offline.

To enable this feature:
 - set enableEnliteBgproxy: true in preferences.json
 - turn on the sensor feature in the medtronic pump.
 - set the sensor id to 0000000
 - select find lost sensor. Glucose should appear after a couple of
cycles.