Add config settings for low-battery shutdown thresholds
Adds a pair of config settings for low-battery shutdown thresholds.
One config setting is for Edison (denominated in millivolts), the
other is for Pi (given as percentage). This config setting is needed
for USB-powered Pi rigs, where the battery level indicator just
returns a random number, which might occasionally be zero.