Bugfixes (#1411)
* Add config settings for low-battery shutdown thresholds

Adds a pair of config settings for low-battery shutdown thresholds.
One config setting is for Edison (denominated in millivolts), the
other is for Pi (given as percentage). This config setting is needed
for USB-powered Pi rigs, where the battery level indicator just
returns a random number, which might occasionally be zero.

* If unable to update with git pull, prompt before aborting

This is mainly for convenience during development; a checkout in
some weird branch doesn't necessarily return success when you
"git pull", but that doesn't necessarily mean you want to abort
oref0-runagain.sh.

* Fix bug that would clobber sshd_config

Depending on the initial contents of sshd_config, this would sometimes
try to add a line to it, but instead completely overwrite it to contain
only that line.

* Extend timeout on JS syntax check unit tests

On Pi Zero hardware, these tests would sometimes time out because
starting the nodejs interpreter is slow (even with a fixed nodejs
interpreter). Extend the timeout to 4s, from the default of 2s, so they
pass.

* Clean up require() usage in IOB unit test, making it pass without timeout on Pi Zero

* Add bash-unit-test-temp to gitignore

* Check for bad (super slow) RPi nodejs versions, install nvm version if found

* Tweak log messages about battery level