Node and npm versions conflict fix during the installation of 0.7.1 openaps
Installing the nodejs 8.x package puts npm version 6.13.4. The installation line of the latest version of npm leads to a conflict between the versions of npm and nodejs. Too new npm cannot run on too old nodejs. So I just removed the npm update to the latest version so that the Openaps installation script would not crash.